requirements.txt
numpy==1.17.2
Pillow==9.4.0
python_pptx==0.6.21
tifffile==2021.11.2
